Extract pattern priority-clipping into shared Rendering.Scene seam
Move the NetTopologySuite pattern-fill priority-clip algorithm out of the
Mapsui renderer and into a new backend-neutral PatternPriorityClipper in
EncDotNet.S100.Rendering.Scene, so a single implementation can serve both
render paths. This is a pure refactor: MapsuiDisplayListRenderer now delegates
to PatternPriorityClipper.Clip and behavior is unchanged.

- Promote NetTopologySuite (2.6.0, already transitive via Mapsui.Nts) to a
  direct managed package referenced by Rendering.Scene.
- Add PatternPriorityClipper (Clip + SimplifyForClip + ExtractPolygonal +
  tolerance/gate constants) with PatternGroup/ClippedPattern value types.
- Delegate MapsuiDisplayListRenderer.ClipPatternsByPriority to the shared
  clipper; retarget the clip-cache key inputs (MapsuiDatasetRenderer) and the
  ExtractPolygonal test to the shared constants/method; refresh cache doc refs.

Copilot-Session: a751ce6b-4361-42aa-a54d-4e06edb4108f
